Melhor distribuição das bombas e melhor posicionamento dos sprites

+++ src/game_objects/Position.java
@@ -1,9 +1,9 @@
 package game_objects;
 
 public class Position {
-	public float x, y;
+	public int x, y;
 	public Position() {}
-	public Position(float x, float y) {
+	public Position(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
